Add the significancy stars
Built on Alex's perfect anova_tukey_tests.R script, add two line (just two) of codes to make the significancy in the out put summary csv file:
If a p-value is less than 0.05 it is flagged with one star (*). If a p-value is less than 0.01 it is flagged with two stars (**). If a p-value is less than 0.001 it is flagged with three stars (***).